requirements.txt
requests==2.28.2
discord.py==2.1.1
python-dotenv==0.21.1
openai==1.6.1